Further subdivided conference list
